ARM: OMAP2+: AM335x: Add BBGW and BBBW support
Both BBGW and BBBW are Boneblack family boards and more or less
same as BBB, with difference in hardware in order to support
onboard WL1835 chip. BBGW does this by removing HDMI receiver
and ethernet PHY, and BBBW removes ethernet PHY, and solder
with a TI WiLink wl1835 chip.

For BBGW, checkout Seeed Studio wiki page for detail:

http://wiki.seeed.cc/

For BBBW, checkout this one on github:

https://github.com/beagleboard/beaglebone-black-wireless

This patch make necessary changes to bring up the WL1835 BT/WIFI
for use with Android, includes:

1. Remove unused pinmux / codes for removed hardware like
   ethernet PHY

2. Setup pinmux for WL1835 and add neccessary initialization
   entry functions

3. Disable lcd_data16 ~ lcd_data23, which has pin conflicts
   with WL1835MOD pins

4. Use PROFILE to differentiate boards: PROFILE_1 for BBB,
   PROFILE_2 for BBGW, and PROFILE_3 for BBBW

Change-Id: I2347fe619c0f770195992ca6dab144d97c794ce5